backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
boto3==1.34.0
python-decouple==3.8
django-storages==1.14.2
channels==4.0.0
channels-redis==4.1.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.7
sendgrid==6.11.0
stripe==7.8.0
requests==2.31.0
pytz==2023.3
